requirements.txt
streamlit==1.29.0
pandas==2.1.0
matplotlib==3.7.2
seaborn==0.12.2
langchain==0.1.0
langchain_experimental==0.0.49
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
openpyxl==3.1.2
langchain_google_genai==0.0.5
tabulate==0.9.0